pda补货bug

+++ warehouse-operation-impl/src/main/java/com/baozun/scm/primservice/whoperation/manager/warehouse/inventory/WhSkuInventoryManagerImpl.java
@@ -8557,48 +8557,19 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
             // 删除待移入
             Double sumQty = 0.0;
             for (WhSkuInventoryTobefilled invTobefilled : invTobefilledList) {
-                if(null != outerContainerId){
-                    if(null != invTobefilled.getOuterContainerId() && outerContainerId.equals(invTobefilled.getOuterContainerId())){
+                    if(null != invTobefilled.getOuterContainerId() && invTobefilled.getOuterContainerId().equals(outerContainerId)){
                         WhSkuInventoryTobefilled cInv = new WhSkuInventoryTobefilled();
                         BeanUtils.copyProperties(invTobefilled, cInv);
                         whSkuInventoryTobefilledDao.deleteByExt(cInv.getId(), ouId);
                         insertGlobalLog(GLOBAL_LOG_DELETE, cInv, ouId, userId, null, null);
-                    }else{
-                        String toBeSkuAttrId = SkuCategoryProvider.getSkuAttrIdByWhSkuInvTobefilled(invTobefilled);
-                        if (skuAttrIds.equals(toBeSkuAttrId)) {
-                            sumQty += invTobefilled.getQty();
-                            Double tobefilledQty = sumQty - invCmd.getOnHandQty(); // 待移入库存还剩下的sku数量
-                            if (tobefilledQty.doubleValue() < 0) {
-                                WhSkuInventoryTobefilled cInv = new WhSkuInventoryTobefilled();
-                                BeanUtils.copyProperties(invTobefilled, cInv);
-                                whSkuInventoryTobefilledDao.deleteByExt(cInv.getId(), ouId);
-                                insertGlobalLog(GLOBAL_LOG_DELETE, cInv, ouId, userId, null, null);
-                                continue;
-                            }
-                            if (tobefilledQty.doubleValue() == 0) {
-                                WhSkuInventoryTobefilled cInv = new WhSkuInventoryTobefilled();
-                                BeanUtils.copyProperties(invTobefilled, cInv);
-                                whSkuInventoryTobefilledDao.deleteByExt(cInv.getId(), ouId);
-                                insertGlobalLog(GLOBAL_LOG_DELETE, cInv, ouId, userId, null, null);
-                                break;
-                            }
-                            if (tobefilledQty.doubleValue() > 0) {
-                                WhSkuInventoryTobefilled cInv = new WhSkuInventoryTobefilled();
-                                BeanUtils.copyProperties(invTobefilled, cInv);
-                                cInv.setQty(tobefilledQty);
-                                whSkuInventoryTobefilledDao.saveOrUpdateByVersion(cInv);
-                                insertGlobalLog(GLOBAL_LOG_UPDATE, cInv, ouId, userId, null, null);
-                                break;
-                            }
-                        }
                     }
-                }else{
                     if(null != invTobefilled.getInsideContainerId() && turnoverBoxId.equals(invTobefilled.getInsideContainerId())){
                         WhSkuInventoryTobefilled cInv = new WhSkuInventoryTobefilled();
                         BeanUtils.copyProperties(invTobefilled, cInv);
                         whSkuInventoryTobefilledDao.deleteByExt(cInv.getId(), ouId);
                         insertGlobalLog(GLOBAL_LOG_DELETE, cInv, ouId, userId, null, null);
-                    }else{
+                    }
+                    if(null == invTobefilled.getOuterContainerId() && null == invTobefilled.getInsideContainerId()){
                         String toBeSkuAttrId = SkuCategoryProvider.getSkuAttrIdByWhSkuInvTobefilled(invTobefilled);
                         if (skuAttrIds.equals(toBeSkuAttrId)) {
                             sumQty += invTobefilled.getQty();
@@ -8626,7 +8597,6 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
                                 break;
                             }
                         }
-                    }
                 }
             }
         }
@@ -9100,7 +9070,7 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
                 Double onHandQty = invCmd.getOnHandQty();
                 if (skuScanQty.doubleValue() == onHandQty.doubleValue()) {
                     // 添加库位库存
-                    this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, skuScanQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, newTurnoverBoxId, isTabbInvTotal, ouId, userId, whSkuAttrId);
+                    this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, skuScanQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, turnoverBoxId, isTabbInvTotal, ouId, userId, whSkuAttrId);
                     // 删除容器库存
                     this.replenishmentUpdateContainerInventory(invCmd, skuScanQty, isTabbInvTotal, ouId, userId);
                     isPutaway = true;
@@ -9110,7 +9080,7 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
                     sum += onHandQty.doubleValue();
                     if (skuScanQty.doubleValue() == sum.doubleValue()) {
                         // 添加库位库存
-                        this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, onHandQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, newTurnoverBoxId, isTabbInvTotal, ouId, userId,
+                        this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, onHandQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, turnoverBoxId, isTabbInvTotal, ouId, userId,
                                 whSkuAttrId);
                         // 删除容器库存
                         this.replenishmentUpdateContainerInventory(invCmd, onHandQty, isTabbInvTotal, ouId, userId);
@@ -9118,7 +9088,7 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
                     }
                     if (skuScanQty.doubleValue() > sum.doubleValue()) {
                         // 添加库位库存
-                        this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, onHandQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, newTurnoverBoxId, isTabbInvTotal, ouId, userId,
+                        this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, onHandQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, turnoverBoxId, isTabbInvTotal, ouId, userId,
                                 whSkuAttrId);
                         // 删除容器库存
                         this.replenishmentUpdateContainerInventory(invCmd, onHandQty, isTabbInvTotal, ouId, userId);
@@ -9127,7 +9097,7 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
                     if (skuScanQty.doubleValue() < sum.doubleValue()) {
                         Double qty = skuScanQty - (sum - onHandQty);
                         // 添加库位库存
-                        this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, qty, locationId, newTurnoverBoxId, isTV, isBM, isVM, newTurnoverBoxId, isTabbInvTotal, ouId, userId, whSkuAttrId);
+                        this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, qty, locationId, newTurnoverBoxId, isTV, isBM, isVM, turnoverBoxId, isTabbInvTotal, ouId, userId, whSkuAttrId);
                         // 删除容器库存
                         this.replenishmentUpdateContainerInventory(invCmd, qty, isTabbInvTotal, ouId, userId);
                         isPutaway = true;
@@ -9136,7 +9106,7 @@ public class WhSkuInventoryManagerImpl extends BaseInventoryManagerImpl implemen
 
                 if (skuScanQty.doubleValue() < onHandQty.doubleValue()) {
                     // 添加库位库存
-                    this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, skuScanQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, newTurnoverBoxId, isTabbInvTotal, ouId, userId, whSkuAttrId);
+                    this.replenishmentSplitPutaway(occupationLineId, occupationCode, cacehSnList, invTobefilledList, invSnCmd, invCmd, skuScanQty, locationId, newTurnoverBoxId, isTV, isBM, isVM, turnoverBoxId, isTabbInvTotal, ouId, userId, whSkuAttrId);
                     // 删除容器库存
                     this.replenishmentUpdateContainerInventory(invCmd, skuScanQty, isTabbInvTotal, ouId, userId);
                     isPutaway = true;
